Ignore files to make empty directories show up
These files are really of no consequence at a later date once some real files are added to these directories. Gitignores simply tell git which files not to track. Since git will only push files and not empty directories this was a workaround to get git to push the empty dirs for now.